requirements.txt
ansible-core>=2.19.3
ansible-pylibssh>=1.3.0
jxmlease>=1.0.3
ncclient>=0.6.19
